requirements.txt
appnope==0.1.0
arrow==0.12.1
async-timeout==4.0.2
atomicwrites==1.3.0
attrs==19.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
chardet==3.0.4
click==6.7
decorator==4.1.2
Deprecated==1.2.13
Django==1.11.20
djangorestframework==3.9.1
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
idna==2.8
imutils==0.5.2
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==2.0.1
meld3==1.0.2
more-itertools==5.0.0
numpy==1.23.4
opencv-contrib-python==4.6.0.66
packaging==21.3
pathlib2==2.3.0
pdfrw==0.4
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.9.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.8.0
Pygments==2.2.0
pyinstrument==3.4.2
pyinstrument-cext==0.2.4
pyparsing==3.0.9
PyPDF2==1.26.0
pytest==4.4.0
python-dateutil==2.7.2
python-slugify==3.0.3
pytz==2017.2
redis==4.3.4
requests==2.21.0
rq==1.11.1
rq-dashboard==0.3.10
scandir==1.10.0
simplegeneric==0.8.1
six==1.11.0
supervisor==4.2.4
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.1
uWSGI==2.0.20
Wand==0.5.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.14.1